+++ src/data/pokemon-species.ts
@@ -7,7 +7,8 @@ import type { GameMode } from "#app/game-mode";
 import { randSeedInt, randSeedGauss } from "#app/utils";
 import type { GrowthRate } from "#enums/growth-rates";
 import type { EvolutionLevel } from "#app/data/balance/pokemon-evolutions/pokemon-evolutions";
-import { pokemonEvolutions, pokemonPrevolutions } from "./balance/pokemon-evolutions/init-pokemon-evolutions";
+import { pokemonEvolutions } from "./balance/pokemon-evolutions/init-pokemon-evolutions";
+import { pokemonPrevolutions } from "./balance/pokemon-evolutions/pokemon-evolutions";
 import type { ElementalType } from "#enums/elemental-type";
 import { variantData } from "#app/data/variant";
 import { SpeciesFormKey } from "#enums/species-form-key";
